Organize functions into separate files
Use more general credentials for sending emails for testing purpoese
